# api/swagger.yml
openapi: 3.0.3
info:
  title: Recommendation System - OpenAPI 3.0
  description: |-
    This is recommendtaion system based on the OpenAPI 3.0 specification.
  version: 1.0.0
...
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /authed/recommendation:
    get:
      tags:
        - recommendation
      summary: Get recommendation
      description: |-
        This can only be done by the logged in user. (Bearer Authentication)
      operationId: getRecommendation
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          description: Current cursor of the page ([Cursor Pagination](https://www.merge.dev/blog/cursor-pagination))
          required: false
          explode: false
          schema:
            type: integer
            example: 5

        - name: pageSize
          in: query
          description: |-
            Max number of products to be returned in the response.

            `Default`: 5
          required: false
          explode: false
          schema:
            type: integer
            example: 5
            minimum: 1
            maximum: 20

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  nextCursor:
                    type: integer
                    description: nextCursor will be found only when there is a next page.
                    example: 5
        '400':
          description: Invalid cursor or pageSize
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Product:
      type: object
      required:
        - productId
        - title
        - price
        - description
        - category
      properties:
        productId:
          type: integer
          example: 10
        title:
          type: string
          example: SanDisk SSD PLUS 1TB Internal SSD - SATA III 6 Gb/s
        price:
          type: number
          format: double
          example: 109
        description:
          type: string
          example: Lorem Ipsum is simply dummy text of the printing and typesetting industry.
        category:
          type: string
          example: electronics
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: This is an example error message.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT